:recycle: Move all chart related code to own moudle 2/3
