+++ libs/designsystem/src/lib/components/charts/shared/index.ts
@@ -1,2 +1,3 @@
 export * from './charts.types';
 export * from './base-chart';
+export * from './base-chart/chart-js/test-utils';
