+++ libs/designsystem/src/lib/components/charts/stock-chart/index.ts
@@ -1,0 +1,1 @@
+export * from './stock-chart.component';
