+++ libs/designsystem/src/lib/components/charts/stock-chart/stock-chart.component.ts
@@ -1,14 +1,9 @@
 import { Component, Input } from '@angular/core';
 
-import { ChartDataLabelOptions } from '../chart';
-import { ChartJSService } from '../chart/chart-js/chart-js.service';
-import { BaseChartComponent } from '../chart/chart.component';
+import { BaseChartComponent, ChartDataLabelOptions, ChartJSService } from '../shared';
 
 import { StockChartJSService } from './stock-chart-js.service';
 
-/* 
-TODO: Find a better way to share decorator settings
-*/
 @Component({
   selector: 'kirby-stock-chart',
   providers: [{ provide: ChartJSService, useClass: StockChartJSService }],
